simplify config & migrate to new SMAPI config helper
