--- TimeSpeed/config.json
@@ -1,16 +1,0 @@
-{
-  "DefaultTickLength": 14.0,
-  "TickLengthByLocation": {
-    "Indoors": 14.0,
-    "Outdoors": 14.0
-  },
-  "EnableOnFestivalDays": false,
-  "FreezeTimeAt": null,
-  "LocationNotify": false,
-  "Keys": {
-    "FreezeTime": "N",
-    "IncreaseTickInterval": "OemPeriod",
-    "DecreaseTickInterval": "OemComma",
-    "ReloadConfig": "B"
-  }
-}
